Extracted Sort select dropdown from student view it's own component file. Reused in team view to sort students for selected staff.
