+++ client/src/pages/Dashboard/views/Students/components/StudentListControls.js
@@ -10,44 +10,9 @@ import { useDashboardDispatch, getDashboardAction as gda } from "pages/Dashboard
 import { EDIT_STUDENT } from "pages/Dashboard/store/actions";
 import Dropdown from "components/Dropdown";
 import { useStudentGroupings } from "pages/Dashboard/utils/student";
+import SortSelectDropdown from "pages/Dashboard/components/SortSelectDropdown";
 
 const { Input } = FormCollection;
-
-const sortTypes = [
-    {
-        key: "name:asc",
-        label: "By name A to Z",
-        icon: "sort-alpha-down"
-    },
-    {
-        key: "name:desc",
-        label: "By name Z to A",
-        icon: "sort-alpha-up-alt"
-    },
-    {
-        key: "priorityLevel:asc",
-        label: "By priority 1 to 10",
-        icon: "sort-numeric-down"
-    },
-    {
-        key: "priorityLevel:desc",
-        label: "By priority 10 to 1",
-        icon: "sort-numeric-up-alt"
-    }
-];
-
-// const groupTypes = [
-//     {
-//         key: "staff",
-//         label: "Group by staff",
-//         icon: "user-friends"
-//     },
-//     {
-//         key: "priority",
-//         label: "Group by priorty",
-//         icon: "exclamation-circle"
-//     }
-// ];
 
 const StudentListControls = ( { sort, groupBy, search } ) => {
 
@@ -58,7 +23,6 @@ const StudentListControls = ( { sort, groupBy, search } ) => {
         icon
     }) );
     const groupLabel = <Icon icon="columns" />
-    const sortLabel = <Icon icon={sortTypes.find( ({key}) => key === sort[0] ).icon} />
 
     return (
         <div className="is-flex mb-5">
@@ -73,20 +37,7 @@ const StudentListControls = ( { sort, groupBy, search } ) => {
                 placeholder="Search Name"
                 onChange={(e) => search[1](e.target.value)} style={{maxWidth:"135px"}}
             />
-            <Dropdown className="ml-2 is-right" label={sortLabel}>
-                {sortTypes.map( sortType => {
-                    const classes = ["dropdown-item"];
-
-                    if( sortType.key === sort[0] ) classes.push("is-active");
-
-                    return (
-                        <Button key={sortType.key} className={classes.join(" ")} size="small" onClick={() => sort[1](sortType.key)}>
-                            <Icon icon={sortType.icon} />
-                            <span>{sortType.label}</span>
-                        </Button>
-                    )
-                })}
-            </Dropdown>
+            <SortSelectDropdown state={sort} className="ml-2 is-right" />
             <Dropdown className="is-right ml-2" label={groupLabel}>
                 {[ { key: "none", label: "No Grouping", icon: "ban" }, ...groupTypes ].map( groupType => {
                     const classes = ["dropdown-item"];
